# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Autoquímicos - Sistema de Gestión de Salas
  description: |
    API REST para el sistema de gestión de salas de conferencias de Autoquímicos S.A. de C.V.
    
    ## Características
    - Gestión de usuarios con roles (Administrador, Aprobador, Organizador, Recepción)
    - Reservas de salas con flujo de aprobación
    - Sistema de coffee break automático
    - Check-in y control de asistencia
    - Auditoría completa de acciones
    - Reportes y estadísticas
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Incluye el token en el header:
    ```
    Authorization: Bearer <token>
    ```
    
  version: 1.0.0
  contact:
    name: Autoquímicos S.A. de C.V.
    email: admin@autoquimicos.com
  license:
    name: Propietario
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://api.autoquimicos.com/api
    description: Servidor de producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id_usuario:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Pérez"
        correo:
          type: string
          format: email
          example: "juan@autoquimicos.com"
        telefono:
          type: string
          example: "555-0123"
        role:
          type: string
          enum: [administrador, aprobador, organizador, recepcion]
          example: "organizador"
        activo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error interno del servidor"
        message:
          type: string
          example: "Descripción detallada del error"
...
paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: Autentica un usuario y devuelve un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@autoquimicos.com"
                password:
                  type: string
                  example: "admin123"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Inicio de sesión exitoso"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    get:
      tags:
        - Autenticación
      summary: Verificar token
      description: Verifica si un token JWT es válido
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  valid:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
...